Fix: turn 타입 int로 변경

+++ BE/src/main/java/com/chibbol/wtz/domain/job/repository/UserAbilityRecordRedisRepository.java
@@ -23,7 +23,7 @@ public class UserAbilityRecordRedisRepository {
 
     public List<UserAbilityRecord> findAllByRoomSeq(Long roomSeq) {
         List<UserAbilityRecord> resultList = new ArrayList<>();
-        String pattern = generateKey(roomSeq, -1L);
+        String pattern = generateKey(roomSeq, -1);
 
         Set<String> keys = redisTemplate.keys(pattern);
 
@@ -38,20 +38,20 @@ public class UserAbilityRecordRedisRepository {
     }
 
     public void deleteAllByRoomSeq(Long roomSeq) {
-        String pattern = generateKey(roomSeq, -1L);
+        String pattern = generateKey(roomSeq, -1);
         Set<String> keys = redisTemplate.keys(pattern);
         if(keys != null) {
             redisTemplate.delete(keys);
         }
     }
 
-    public List<UserAbilityRecord> findAllByRoomSeqAndTurn(Long roomSeq, Long turn) {
+    public List<UserAbilityRecord> findAllByRoomSeqAndTurn(Long roomSeq, int turn) {
         String key = generateKey(roomSeq, turn);
         List<Object> jsonDataList = redisTemplate.opsForHash().values(key);
         return convertJsonDataListToUserAbilityRecordList(jsonDataList);
     }
 
-    public UserAbilityRecord findByRoomSeqAndTurnAndUserSeq(Long roomSeq, Long turn, Long userSeq) {
+    public UserAbilityRecord findByRoomSeqAndTurnAndUserSeq(Long roomSeq, int turn, Long userSeq) {
         String key = generateKey(roomSeq, turn);
         String userSeqField = userSeq.toString();
         String jsonData = (String) redisTemplate.opsForHash().get(key, userSeqField);
@@ -93,7 +93,7 @@ public class UserAbilityRecordRedisRepository {
         });
     }
 
-    private String generateKey(Long roomSeq, Long turn) {
+    private String generateKey(Long roomSeq, int turn) {
         return turn == -1L ? KEY_PREFIX + ":room:" + roomSeq + ":turn:*" : KEY_PREFIX + ":room:" + roomSeq + ":turn:" + turn;
     }
